feat: Sell content of ShippingBinTile at 5pm
TimeManager alerts all instances of registered TimeManagerListener at the top of each new hour.

SceneFarm implements TimeManagerListener and checks if the hourly alerts is 5pm. If so, it totals and sells the stash of Sellable instances, increments the player's currencyNuggets field (which is being rendered by the HUD [beneath the image of the player's current item]), and clears ShippingBinTile's stash of Sellable objects.