fix: Check for entity-collision before dropping Holdable
Stumbled into another issue (but fixed it): dropping Holdable was functional when the player was facing RIGHT and DOWN, but wasn't for LEFT and UP. The reason for this was entity-collision checking had counted the player's holdable (positioned in the top-left quadrant of the player) as a positive collision (which prevented the dropping method to be called).

NEXT: Implement Holdable for Bush and Rock.