refactor: Add default implementation to Scene.getInputButtonPad()
Subclasses of Scene can override this "template" or they can override the individual methods (e.g. doButtonJustPressedA(), doButtonJustPressedB(), and doButtonJustPressedMenu()) called from this "template".

PURPOSE: The intent of this commit is to allow all of PoohFarmerCartridge's scenes to (by default) inherit the picking-up-and-dropping-of-Holdable-objects functionality. Scenes from PocketCrittersCartridge are mainly overriding doButtonJustPressedA().