refactor: Break ProductEntity when dropped
ProductEntity's isSolid is changed to false and the image is updated to its broken version when drop(Tile) is called. ProductEntity.isSolid being false will also cause timeLeftBeforeRemoval to decrement within update(long). When timeLeftBeforeRemoval reaches 0, the ProductEntity's active is set to false and the scene's EntityManager will remove it from the game.