tinker: Add EggEntity that can change SceneChickenCoop's isEggIncubating
EggEntity can render a broken version of itself if dropped onto the ground (it is removed from the game after 2 seconds). I messed up the alpha channel of the sprite sheet containing the shipping bin when adjusting the image of the egg's broken version.

SceneChickenCoop's isEggIncubating currently controls rendering an IMAGE of an egg in the proper position. This approach was taken (instead of simply using the EggEntity's rendering method) to PREVENT the player from picking up the incubating egg.

NEXT: Use "isEggIncubating" to track the number of days left before the egg hatches into a baby chicken.