feat: Establish feeding system for chicken
Collect adult chicken instances, sort them based on daysUnhappy and daysAlive, determine whether enough fodder in feeding area for ALL of the adult chickens (ALL chicken LAY AN EGG IN A RANDOM POSITION and decrement daysUnhappy) or NOT ENOUGH fodder for ALL of the adult chickens (SOME chicken LAY AN EGG IN A RANDOM POSITION and decrement daysUnhappy; SOME chicken call their becomeUnhappyDueToMissedFeeding() method [for 3 days, this chicken cannot lay eggs but should still be fed in order to decrease daysUnhappy]).

FIX: Can no longer drop an EggEntity onto the EggIncubatorTile when SceneChickenCoop.isEggIncubating is true (can't put an egg in the incubator if it's already occupied).