tinker: Update CountZeroActivity to use Switch widgets
Previous implementation used a ListView with an adapter that defined how each list-item is displayed (in this case, selecting an image for an ImageView based on each entry's isActive field).

Current implementation uses a ScrollView with one child (a LinearLayout loaded with Switch Views [widgets]). It's also using October dates.

SUMMARY: Experimenting with a new layout component.