refactor: Use List<Dvd> instead of Dvd[]
Using new solution to resolve the issue of marshalling JSON to a parameterized collection of objects (type erasure when deserializing). Instead of using RestTemplate's getForObject() to return an array of Dvd, it's now using exchange() and passing in a ParameterizedTypeReference as an argument.

Updated AdapterRecyclerView to take a List<Dvd> instead of a Dvd[].

Previously, I was creating a new adapter for the RecyclerView every time I updated the dataset. Now it's being handled more properly, using the same dataset reference and calling the adapter's "notifyDataSetChanged()".