feat: Add modal bottom sheet by extending "BottomSheetDialogFragment"
Included "Material Components for Android" library in the project's list of dependencies.

Clicking an item from the RecyclerView will trigger the bottom sheet to show. Currently, not passing data from the item clicked to the bottom sheet (clicking any dvd item will show the same, uncustomized bottom sheet dialog). Clicking outside of the bottom sheet or sliding it downward will cause it to be dismissed.