tinker: Replace bottom sheet with an alert dialog
Selecting an item from the RecyclerView should display more information about the selected item (e.g. its title and availability), and allow for user to add it to a "checkout-cart". Will also need a way to allow user to view the content of the "checkout-cart", as well as remove items from it.

Will also need a way for user to actually checkout the selected items (updating the database).

Maybe allow access to view cart, remove item from cart, and checkout content of cart through the bottom sheet (one of the RecyclerViewActivity's buttons can trigger showing the bottom sheet).