refactor: Populating bottom sheet's ListView with List<Command>
Previously, loaded the ListView using a String array stored in the resource file strings.xml.

Now generating a List of Command objects in RecyclerViewActivity and supplying that to CommandToListItemAdapter. The "menu items" of the bottom sheet will now call the selected Command object's execute() when they get clicked.

Next, implement the cart feature (which should lead to actually checking out the dvd(s) [updating the database]).