feat: Implement Listener/Observer pattern for DisplayDvdDialogFragment
Now RecyclerViewActivity can respond to the user selecting "Add to cart" or "Cancel" from the DisplayDvdDialogFragment.

RecyclerViewActivity has a cart. The ViewContentOfCartCommand is only displaying a Toast-message to verify that content really was added to the cart.

NEXT: Should ViewContentOfCartCommand handle checking-out/removing-item-from-cart by: launching a new Activity, swapping Fragments, or handling it through a Dialog?