feat: Implement viewing content of cart
Launching a new Activity (ViewCartActivity) has data attached to its Intent (the list of Dvd instances in the user's cart). ViewCartActivity uses this data to populate its RecyclerView.

Presently, outputting Toast-messages when RecyclerView's items are clicked (to become a launcher of a remove-dvd-from-cart confirmation dialog). Outputting a Toast-message when the check out button is clicked.