refactor: Change to Fragment-oriented app design
Previously, a RecyclerView was statically defined in the Activity's layout XML file. Now, a FrameLayout is statically defined in the layout file and acts as a placeholder/Fragment-container. A Fragment with a RecyclerView is dynamically added to the Activity's FrameLayout via FragmentTransaction.

The initialization code for the RecyclerView was moved from "onCreate()" to "onStart()" because of null pointer issues.

NEXT: Perform FragmentTransaction when calling SearchByTitleCommand.execute(). This should replace the RecyclerView with a form-like fragment (to allow the user to specify a title) that can add a new Dvd to the database/repository.