tinker: Add PassingThroughActivity
An updated version of [JackInActivity].

Better use of fragments and dialogs (avoids having to save everything, change to a new activity, return from that activity, reload the game with returned data).

Head-up-display is separated from the game's viewport.

Diagonal movement and diagonal tile/item/entity checking.

For the most part, it was much cleaner and more organized than [JackInActivity]... until bringing in SceneManager and its sceneStack.

Bugs are present related to saving/loading of scenes on the sceneStack (and scenes not on the sceneStack that have their xLastKnown and yLastKnown specified) with GameCamera's widthViewport/heightViewport values.