fix: Resolve issue with simultaneous touch input
There was a sequence-dependent bug related to picking up item with a-button (button pad) while moving the player (holding down direction pad). The player was able to pick up items if a-button is held down first, and moving occurred second. However, if a direction pad button was held down first, and a-button was pressed afterward, the item would not be picked up.

The problem was found in InputManager.onDirectionPadTouch(), where I was setting the previous values of movement buttons to false [resetting them]. I was accidentally resetting ALL buttons (instead of just movement related buttons), so button presses from button pad getting turned off.