feat: Remove game loop's "work simulator" (Thread.sleep())
Moving was a little choppy so I decided to see if removing the "work simulator" would help. It did.

Rendering tiles in the background suddenly became really smooth. It had such a significant effect that I had to reduce moveSpeed from 4f to 1f.

Also, renamed listener related components for DirectionPadFragment and ButtonPadFragment.