feat: Add save/load by user through menu state
MenuStateImpl.interpretInput() (which is being called from the game loop thread) is where there is a check for user specified "load". The steps to "load" involves manipulating the source image of ButtonHolder from the StatsDisplayerFragment's View hierarchy (from the main UI thread). It was necessary to call "loadFromFile()" from the UI thread. However, this brought another issue. Loading is happening on the main thread while the game loop thread is simutaneously continuing on with its update-render-update-render-update-render-etc loop. Since "load" from MenuStateImpl is happening during update, the game loop will then try to render (but loading is happening simutaneously... which led to null pointer exceptions). It was necessary to wait for the main UI thread to finish its call to "load" before allowing the game loop thread to continue its update-render-update-render-etc cycles. I found a solution on StackOverflow that uses the FutureTask class's "get()" to block the game loop thread until the task of loading is finished on the main UI thread.

Also, MenuStateImpl is a much cleaner version of StartMenuState (from JackInActivity).