diagnose: Discover issue with transfer point shrinking
Transfer points collision Rect appear to be flattening relative to the direction in which the player collided with it (e.g. the player moving upward/downward during the collision would lead to the transfer point's Rect being squished vertically... the player moving leftward/rightward during the collision would lead to the transfer point's Rect being squished horizontally).

When multiple transfer points are present in the scene, only the collided Rect is affected; the other transfer points' Rect are still their full size).