tinker: Add HeadUpDisplay to SceneEvo
SeaJelly is able to damage the player. When SeaJelly moves, it may trigger entity collision between itself and the Player. The collision causes Seajelly to go into ATTACk state (changes which animation is used) and damages the player.

The damage is rendered to the screen in red font and gets removed by HeadUpDisplayer after 5 seconds.