fix: Handle Eel.State.ATTACK rendering bug
Eel's attack animation is narrow and tall (unlike Eel's other State's animations, which are wide and short). Previously, when an eel attacked, its narrow and tall image was stretched wide and flattened vertically; it was a very noticeable visual glitch.

Now Eel has an overridden render(Canvas) which transforms attack animation (aligns the image to the left or right, depending on which way it was moving when it collided with the player).

This approach is just to remedy the visual glitch, it does not adjust the actual Entity.bounds which could led to bug related to collision detection when the Eel is in ATTACK state.