fix: Make PassingThroughActivity full-screen
Translating viewport coordinates to in-game coordinates appears to be working. I think the problem was the action bar (its height wasn't factored in relative to the source of the MotionEvent's coordinate system). The action bar is no longer present when this Activity is in full-screen mode, and now the triggering of Entity.die() via the viewport works.

NEXT: decide to continue on with full-screen or keep the action bar, subtract its height out within the calculation for GameCamera.convertToInGameRect(Rect).