refactor: Revert PassingThroughActivity to have action bar
SceneEvo.interpretViewportInput() now account for the height of the action bar in its calculation to define the bounding rectangle of the user's touch.

SUCCESSFULLY converting from screenRect coordinates to in-game coordinates and vice-versa! Viewport is MUCH MORE powerful now and this opens up more fun ways to interact with the game!