feat: Implement Eel State.CHASE functionality
Still experiencing issues with eel changing its size and bounds when going into and coming out of State.ATTACK (its animation is narrow and tall while the other states' animation are wide and short). It's attacking the player properly from all 4 directions (issue with attacking downward was resolved by moving the eel's vertical position upward to avoid overlapping with the player's bounds).

The potential future issues are when the eel is attacking the player while other entities are nearby. The eel's shifting size and bounds may accidentally overlap with a nearby entity's bounds. Maybe do a collision check prior to changing size and bounds.