refactor: Simplify MenuItemInitial's border
Now using the same Rect instance from the background panel, but with a different Paint instance (one that uses Style.STROKE... like from MenuItemEvolution's yellow border cursor). The border is now being rendered above the background panel, both are the exact same size (with the exact same roundness for their corners).

The previous implementation for rendering the border was to create a slightly larger Rect than the background panel, render it first, and then render the background panel over it. This worked, but the calculations to create a slightly larger Rect was messy. It also led to uneven roundness of the corners (the background panel's corners were slightly different from the border's corners).