feat: Add "SHAKING" state to GameCamera
Currently shaking for 5 seconds when a kelp instance is clicked using the viewport. 5 seconds is way too long. Also, the shaking sometime moves the GameCamera too far (the player is out of the clip/frame).

The buffed up GameCamera and viewport (translating touch screen coordinates to the in-game coordinate system... knowing that the on-screen click is colliding with the kelp's bounds) is going to make gameplay more fun.