bug: Viewport triggering entity collision is inaccurate when moving
When player is not moving, clicking on kelp from different position in the scene will trigger GameCamera shaking. However, when pressing the direction pad... clicking on kelp no longer trigger GameCamera shaking (even if player isn't really moving in the game [e.g. walking into something solid... in this case the Meat item]). SceneEvo.interpretViewportInput() is outputting a negative y0Event value (pre-converting to in-game coordinates) when the viewport is pressed when there is simutaneous direction pad input.