fix: Account for status bar and action bar height
View.getLocationOnScreen(int[] outLocation) clued me in that my previous implementation of SceneEvo.interpretViewportInput() (where I only factored in the height of the action bar) was only ALMOST correct. I tried to determine why there was a difference between MySurfaceView's y and the height of the action bar. The height of the status bar (in pixel) was a perfect match for this difference.

BUG: still haven't figure out the bug related to pressing the viewport while simutaneously moving. I think it has to do with using a coordinate system relative to the View that received the first touch event (future pointers' position is based on that View's coordinate system).