fix: Record x, y of MotionEvent in InputManager
Previously, was interpreting the MotionEvent's data inside SceneEvo.interpretViewportInput() (which gets called by SceneEvo.update(long)). Now, values of the position of the pointer related to MySurfaceView touch events are recorded in InputManager. SceneEvo.interpretViewportInput() checks the recorded position (relative to MySurfaceView coordinate system) instead of trying to figure out if the event has multiple pointers or not.

The game loop (1) update inputManager, (2) update game, and (3) render game. I think the resolution is related to recording the values before direction-pad input is checked inside the game's update().

This approach fixes the simultaneous direction-pad-and-viewport-touch bug, and also doesn't need to account for the height of the action bar or the status bar.