feat: Add buying functionality to seed shop
Item now has a [price] field. It gets initialize to -1. Concrete subclasses of Item can reassign [price] in their constructor. The seed shop checks that the item's [price] is above zero before performing the transaction. Then there's a check to see if the player has enough [currency] to trade for the item. Then the item is moved to the backpack, currency is decremented, the item is removed from the seed shop's inventory, and the seed shop's adapter for its RecyclerView is notified of the change to its data set.