tinker: 2 solutions to orientation change while Dialog is open
The first (SeedShopDialogFragment) calls dimiss() in onPause() and sets a flag to call show() when reloading. The second (ComputerDialogFragment) is the one I prefer more, it's fetched from the fragment manager via its TAG and has its reference to Game updated to the new Game object.

Handling screen orientation change while a Dialog is open is important if I want to lend into using android widgets.