fix: Refreshing backpack and stats displayer now function as expected
After reloading, items in ButtonHolderA and ButtonHolderB were showing in the backpack dialog (its RecyclerView's adapter wasn't being notified).

There was a recent change to have the game start off with [currency] (instead of starting with the value of zero) in order to test buying functionality. The stats displayer wasn't updated to reflect the new initial [currency] value, now it gets updated near the end of Game.init().