feat: Add editting file name functionality
One of the persisted SavedEntry instance contained a typo (I had two "punch-left" when one of them should've been "punch-right"). Needed a way to update the file name, so I added an edit-file-name-of-saved-entry dialog. Typo's fixed.

BUG: after clicking the "Ok" button, the recyclerview item's layout no longer display all the SavedEntry (it's just the first one [stretched out] and you have to scroll to see the rest).

It'd be cool to be able to play the animation sequence from the saved-list dialog.