fix: Use correct adapter for saved-list dialog
Previously, was using the adapter for the staging area when I needed to use the adapter for the repository (the one for the repository inflates a layout that displays indexColumn and indexRow values, while the one for the staging area does not).

ALSO: Add AutoPilot:Off activity (accessed via MainActivity's options menu).