tinker: Add action bar subtitle and display device's API level
ActionBar was introduced in API level 11. My app's minSdkVersion is 19, so there isn't a compatibility issue (so using a conditional statement to check the device's version of Android and the @TargetApi(11) annotation were unnessary [but good for learning/practice]). CheatActivity gets the device's build version and displays it in a TextView at runtime.