tinker: Add landscape-orientation layout for CrimeFragment
The crime-date button and solved checkbox are stacked above/below each other in the portrait-orientation layout, but are combined into one line in the landscape-orientation layout.

Determining how much space each widget will get on the shared line is based solely on the layout parameter "weight", instead of a combination of the layout parameters "width" and "weight".

The crime-date displayed to the screen uses the DateFormat class to translate the Date object into a simpler format.