tinker: Change AutoPilotOffActivity to derive from SingleFragmentActivity and host a ListFragment
Using an abstract class for all activities that only need one fragment container. This facilitates the re-use of the layout file with the single FrameLayout (now named activity_fragment.xml).

AutoPilotOffActivity now starts off by displaying a collection of Crime instances in a ListView (from the ListFragment), instead of starting off with the detail-fragment.