tinker: Implement CrimeListFragment.onListItemClick(ListView, View, int, long)
Clicking an item from the list-fragment will start an activity hosting the detail-fragment (editting the detail-fragment changes the model instance, and pressing the devices back-button [restarting/resuming the list-fragment] notifies the adapter of changes to its data set [refreshes the items in the ListView]).

Data is passed from the list-fragment to the activity hosting the detail-fragment by putting an extra on the intent that starts the activity. The hosting activity will then call the detail-fragment's factory method (instead of its no-arg constructor) to attach the data-from-the-extra as arguments to the fragment. The fragment arguments are stashed during recreation (due to configuration changes [screen rotation]).