tinker: Use ViewPager widget to display CrimeFragment
Replace CrimeActivity with CrimePagerActivity. Able to navigate through the collection of Crime instances using swipe-left and swipe-right gestures while in the detail-fragment (instead of needing to press the back-button to return to the list-fragment, and selecting another instance of Crime).

Using a new xml resource file (ids.xml) to create a standalone resource ID (since we're instantiating the ViewPager programmatically instead of in a layout file where we can create a resource ID via ViewPager id attribute). "The FragmentManager requires that any view used as a fragment container must have an ID. ViewPager is a fragment container, so you will need to assign an ID to it."