tinker: Retain HelloMoonFragment to prevent interruption of audio playback during configuration change
The fragment's view will be destroyed during screen rotation, but the fragment itself is not (it enters a "retained" state for a very brief period before the new activity's new FragmentManager finds it and recreates its view).

"When you retain a fragment, you can count on all of its instance variables to keep the same values. When you reach for them, they are simply there."