tinker: Add menu item to toggle action bar's subtitle
CrimeListFragment uses an alternative menu file if the device's OS version is HONEYCOMB or above. The alternative menu resides in a resource folder appended with "-v11" and has an extra options menu item that toggles the Activity's subtitle. The state of the subtitle (visible vs hidden) survives rotation, back-button press, and up-button (ancestral navigation via "NavUtils") press.

The last one (up-button press) was tricky. It was solved by declaring a launch mode (with a value of "singleTop") to the PARENT activity in the manifest file (so the system doesn't create a new instance of the activity in the target task).