tinker: Add "empty view" for CrimeListFragment's ListView
Tying into the automatic switching feature (the "empty view" is shown when the ListView is empty, and the list is shown when the ListView has items) that AdapterView (and its subclasses [i.e. ListView]) supports.

The "empty view" displays a message (informing the user that the ListView is empty) and a button (to trigger the creation of a new crime so the user doesn't have to go to the options menu or action bar).