tinker: Save list of crimes to the filesystem
CriminalIntentJSONSerializer converts the list of Crime instances into a JSONArray and writes out that JSONArray to the specified file. Update Crime class to be able to save individual instances of Crime in JSON format (as JSONObject). CrimeLab delegates saving its list of crimes to its instance of CriminalIntentJSONSerializer. The actual call to save the list of crimes occurs in CrimeFragment.onPause().