tinker: Load Crime instances from file
Data now persists between runs of the app. CrimeLab uses a new ArrayList<Crime> if the app is starting for the first time (a file has not been written to the filesystem), and will load (read data from a saved file and convert it from JSONObject to Crime instances) the previous entries if it's not the first run.