tinker: Solve "Context Menus and Contextual Action Mode" chapter's challenge
Implemented a way to delete crime in CrimeFragment (the detail-view [as oppose to the list-view]). Since we're not deleting list item(s) within the context of a list, we do not need a floating context menu or contextual action mode. Being able to delete the crime being displayed in CrimeFragment (detail-view) means the deleting will apply to the screen as a whole, and it would be more appropriate to add the deleting feature in the options menu or on the action bar (I added it to the action bar as an action icon).