tinker: Use two-pane (list-detail) interface for larger screens
Suppose to be an alternative layout for tablet devices, but this is implemented to be seen on phones in landscape orientation (the difference is the directory [default vs ones with screen dimension qualifiers] of the [alias resource] file). An [alias resource], a resource that points to another resource and is conventionally defined in [res/values/refs.xml], was used to reference different layouts using one reference [e.g. R.layout.activity_listdetail].

Fragment callback interfaces were defined in [CrimeListFragment] and [CrimeFragment] and implemented in [AutoPilotOffActivity] to communicate/update data between the list-view and the detail-view when using the two-pane layout. This updating wasn't necessary with the single-pane layout (for phone devices in portrait orientation) because the detail-view completely obscured the list-view and only needed to be updated when resuming the list-view.