tinker: Add RemoteControlActivity (Ch. 24 Styles and Includes)
Learned to define style resources. Styles can be referenced by multiple Views. When a change is needed, instead of adding the change to each of the multiple Views, we only need to add the change to one place (the style being referenced). Styles can be extended (adding new attributes and/or overriding parent attributes).

First time using TableLayout and TableRow.

Learned about the [include] tag and its relationship to the [merge] tag (optimizing the layout hierarchy by not including redundant ViewGroups).