fix: Remove visual artifact and don't define padding
Changed the left and top border values for all three 9-patch images to remove visual artifact (provided extra space at the start and end of the stretchable regions).

Also, if the View that's using the nine patch drawable has a padding defined, it may be overriding the right and bottom border values of the 9-patch image. I removed the padding attribute from RemoteControlFragment's TableLayout, and now it looks like the right and bottom border values from the 9-patch image are being applied.