tinker: Download photos on dedicated background thread and display them on main thread (Ch. 27 Loopers, Handlers, and HandlerThread)
Using a dedicated background thread to download photos. Posting messages to the end of a looper's message queue using a handler. When the looper gets to a particular message in the queue, it gives the message to the message's target to handle (download bytes from the URL and turn those bytes into a bitmap).

Scheduling work on the main thread (updating the UI by setting the bitmap on an ImageView) from the dedicated background thread. Pass in a handler that was created from the main thread into the dedicated background thread. It will work with the looper (and message queue) of the main thread.

Clean up is needed within "onDestroyView()" for screen rotation, in case ThumbnailDownloader hang on to invalid ImageViews.