tinker: Respond to MotionEvent in a custom View (Ch. 32 Custom Views and Touch Events)
Add access to DragAndDrawActivity via AutoPilotOffActivity's options menu. DragAndDrawFragment is compose of a single custom View that interprets touch events as the starting point of a new rectangle, the resizing of the currently-being-drawn rectangle, and the ending  point of the currently-being-drawn rectangle. The custom View creates a Box instance from the values of the starting and ending points, and stores the Box in an ArrayList. The ArrayList is saved and restored during screen orientation changes.

NOTE: View.onSaveInstanceState() and View.onRestoreInstanceState(Parcelable) will only be called if your View has an id.